+++ packages/release-manager/src/index.ts
@@ -1,5 +1,5 @@
 export * from "./types.js";
-export * from "./variables.js";
+export * from "./variables/variables.js";
 export * from "./manager.js";
 export * from "./releases.js";
 export * from "./repositories/release-repository.js";
