+++ packages/release-manager/src/repositories/types.ts
@@ -1,0 +1,9 @@
+import type { Release, ReleaseIdentifier, ReleaseWithId } from "../types.js";
+
+export interface ReleaseRepository {
+  getLatestRelease(options: ReleaseIdentifier): Promise<ReleaseWithId | null>;
+  createRelease(release: Release): Promise<ReleaseWithId>;
+  setDesiredRelease(
+    options: ReleaseIdentifier & { desiredReleaseId: string },
+  ): Promise<any>;
+}
