+++ packages/release-manager/src/variables/variables.ts
